fixed deal method
Deleted the canPlayOn method since i already have a canTakeCards method. Swapped the places of suit and rank since the card's name was printing wrong in the toString method. Added the removeCardFromDeck method to update and remove cards from the deck as the deal method gets called, card indexes should be replaced. Deleted the playCard method, will replace this method with loops in the main program.