added GameState and Pisti class
The GameState class has been created due to the fact that we need a class that will keep track of all the cards that players will acquire and the scoring for each player. This class has also getters for each player score and a getter for determining the winner and returning it. Fixed the error where when a card was played on board the updated deck was also deleting the remaining cards in the deck. Added a Pisti class to keep the game going on and the loops for adding removing and getting cards from the board respectively.