added score class
CanTakeCards method is fixed, Pişti methods are now deleted and instead i created int values to hold counter for each player. I now use those integers in the ScoreGame method for each pişti made by each player. The game is now working just fine except when there is a pişti case it can't acknowledge it and instead print "The same rank played.". This is technically true and nothing is wrong with that but  a specific text should be printed in a Pişti case so, i will work on that later. Other game mechanics and calculating the scores etc. work perfectly.